Add files via upload
This project was created for Tour loving people where a admin can add packages and designed website for motivate people to travel more. And  users can choose the packages as per their requirements. Skills has used like HTML, CSS, Bootstrap, JavaScript as a frontend, PHP as a backend and MySQL for the database.